refactor : Lotto  패키지 수정
