+++ src/main/java/lotto/lotto/LottoEnums.java
