+++ src/main/java/lotto/lotto/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.user;
+package lotto;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.LinkedList;
